Delete a bunch of Kotlin, Swing chat code.
