+++ src/main/kotlin/com/sourcegraph/cody/agent/CodyAgentServer.kt
@@ -191,9 +191,6 @@ interface _LegacyAgentServer {
 
   @JsonRequest("chat/export") fun chatExport(): CompletableFuture<List<ChatHistoryResponse>>
 
-  @JsonRequest("chat/restore")
-  fun chatRestore(params: ChatRestoreParams): CompletableFuture<ConnectionId>
-
   @JsonRequest("attribution/search")
   fun attributionSearch(
       params: AttributionSearchParams
