+++ src/main/kotlin/com/sourcegraph/cody/agent/CodyAgentService.kt
@@ -11,7 +11,6 @@ import com.intellij.openapi.progress.ProcessCanceledException
 import com.intellij.openapi.project.Project
 import com.intellij.util.net.HttpConfigurable
 import com.sourcegraph.cody.agent.protocol.ProtocolTextDocument
-import com.sourcegraph.cody.chat.AgentChatSessionService
 import com.sourcegraph.cody.config.CodyApplicationSettings
 import com.sourcegraph.cody.context.RemoteRepoSearcher
 import com.sourcegraph.cody.edit.EditService
@@ -116,7 +115,6 @@ class CodyAgentService(private val project: Project) : Disposable {
       }
 
       if (!project.isDisposed) {
-        AgentChatSessionService.getInstance(project).restoreAllSessions(agent)
         CodyFileEditorListener.registerAllOpenedFiles(project, agent)
       }
     }
